Implementato ErrorHandler

+++ commands/vote.js
@@ -4,36 +4,22 @@ module.exports = {
 	execute(message, args, moderatore){
 		const embed = require("../functions/sendEmbed.js");
 		const check = require("../functions/checkVotes.js");
-		
-		let mentionedArray= message.mentions.members.array();
+		let err = require("../functions/errors");
+
+		if(err.errors([0, 1, 2, 3, 4, 5], moderatore, message))return;
+
 
 		if(message.channel.name.valueOf() !== message.member.user.id.valueOf()){
 			message.delete();
 			embed.sendEmbed([255, 0, 0], "Hai mandato il voto nella chat sbagliata!", message.channel);
 			return;
 		}
-
-		if(mentionedArray.length !== 1){//if they call more or less then 1 people
-			embed.sendEmbed([255,0,0], "Citare una persona da votare", message.channel);
-			return;
-		}
-
-		if(!moderatore.playerList.has(message.member)){
-			embed.sendEmbed([255,0,0], "Devi essere in gioco per poter eseguire i comandi.", message.channel);
-			return;
-		}
-		
-		let mentioned = message.mentions.members.first();
-
-		if(!moderatore.playerList.has(mentioned)){
-			embed.sendEmbed([255,0,0], "Citare una persona in gioco.", message.channel);
-			return;
-		}
-
 		if(moderatore.ballottaggio.length > 0 && !moderatore.ballottaggio.includes(mentioned)){
 			embed.sendEmbed([255, 0, 0], "Hai votato una persona che non è in ballottaggio!", message.channel);
-			return;
+			return true;
 		}
+
+		let mentioned = message.mentions.members.first();
 
 		moderatore.playerList.get(mentioned).votes.push(message.member);
 		moderatore.numberOfVotes += 1;
@@ -41,6 +27,8 @@ module.exports = {
 		
 
 		if(moderatore.numberOfVotes === moderatore.playerList.size - moderatore.numberOfDeadPlayer - moderatore.ballottaggio.length){
+			//unmuting people
+			unMute.execute(message);
 			check.execute(message,moderatore);
 		}
 		message.channel.delete();
